Refact: ParcelInformation 클래스 위치 수정
올바르지 않은 위치에 있어서 수정했습니다.
