+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Models/ParcelInformation.swift
@@ -1,0 +1,18 @@
+//
+//  ParcelInformation.swift
+//  ParcelInvoiceMaker
+//
+//  Created by JunHeeJo on 1/26/24.
+//
+
+import Foundation
+
+class ParcelInformation {
+    let receiverInfomation: ReceiverInfomation
+    let costInfomation: CostInfomation
+
+    init(receiverInfomation: ReceiverInfomation, costInfomation: CostInfomation) {
+        self.receiverInfomation = receiverInfomation
+        self.costInfomation = costInfomation
+    }
+}
