+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Models/ParcelProcessor.swift
@@ -6,17 +6,6 @@
 
 import Foundation
 
-class ParcelInformation {
-    let receiverInfomation: ReceiverInfomation
-    let costInfomation: CostInfomation
-
-    init(receiverInfomation: ReceiverInfomation, costInfomation: CostInfomation) {
-        self.receiverInfomation = receiverInfomation
-        self.costInfomation = costInfomation
-    }
-}
-
-
 class ParcelOrderProcessor {
     private let database: ParcelInformationPersistence
